+++ src/pages/gd/Claim/ClaimBalance.tsx
@@ -1,0 +1,68 @@
+import React, { useMemo, useCallback } from 'react'
+import { View, Box, Text } from 'native-base'
+import { ArrowButton, BalanceGD } from '@gooddollar/good-design'
+import { useSetChain } from '@web3-onboard/react'
+import { SupportedChains } from '@gooddollar/web3sdk-v2'
+import { ChainIdHex } from '../../../constants'
+import { Fraction } from '@uniswap/sdk-core'
+
+const NextClaim = ({ time }: { time: string }) => {
+    return (
+        <Text fontFamily="subheading" fontWeight="normal" fontSize="xs" color="main">
+            Claim cycle restart every day at {time}
+        </Text>
+    )
+}
+
+const ClaimTimer = () => {
+    const timer = '21:00:00'
+    return (
+        <Box height="50" justifyContent="center" flexDirection="column" my="4">
+            <Text fontFamily="subheading" fontSize="md" color="main">
+                Your next claim
+            </Text>
+            <Text>{timer}</Text>
+        </Box>
+    )
+}
+
+export const ClaimBalance = ({ time, price }: { time: string | false; price: Fraction | undefined }) => {
+    const [{ connectedChain }, setChain] = useSetChain()
+
+    const network = useMemo(
+        () => (connectedChain && connectedChain.id === '0x7a' ? SupportedChains[42220] : SupportedChains[122]),
+        [connectedChain]
+    )
+
+    const switchChain = useCallback(() => {
+        const chain = ChainIdHex[SupportedChains[network as keyof typeof SupportedChains]]
+        void setChain({ chainId: chain })
+    }, [setChain])
+
+    return (
+        <View textAlign="center" display="flex" justifyContent="center" flexDirection="column" w="full" mb="4">
+            <Box backgroundColor="goodWhite.100" borderRadius="15" p="1" w="full" h="34" justifyContent="center">
+                <NextClaim time={time || ''} />
+            </Box>
+
+            <ClaimTimer />
+            <Box borderWidth="1" borderColor="borderGrey" width="90%" alignSelf="center" my="2" />
+            <Box>
+                <BalanceGD gdPrice={price} />
+            </Box>
+            <Box alignItems="center">
+                <ArrowButton
+                    borderWidth="1"
+                    borderColor="borderBlue"
+                    px="6px"
+                    width="200"
+                    text={`Claim on ${network}`}
+                    onPress={switchChain}
+                    innerText={{
+                        fontSize: 'sm',
+                    }}
+                />
+            </Box>
+        </View>
+    )
+}
